services/guidance/pyproject.toml
[project]
name = "gym-guidance"
version = "0.1.0"
description = "LLM guidance service — form alert handler, conversation manager, notification dispatcher"
requires-python = ">=3.11"
dependencies = [
    "gym-shared",
    "anthropic>=0.30",
    "google-genai>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guidance"]

[tool.uv.sources]
gym-shared = { workspace = true }
